+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## [19.9.0]
 - Bottom sheets can now be set to not be dismissed when user swipes down the bottomsheet.
-    - Added two new properties to BottomSheet: 'IsClosableOnSwipe' and 'OnBackButtonPressedCommand'.g
+    - Added two new properties to BottomSheet: 'IsInteractiveCloseable' and 'OnBackButtonPressedCommand'.
 
 ## [19.8.0]
 - Added `TextColor` and `Style` properties to `SubtitleOptions`